#113 test: マスタ系の画面のテストを抽象化する
